jsdialog: change sendDialogEvent API parameter
WindowId with 64bit length is needed

Reviewed-on: https://gerrit.libreoffice.org/c/core/+/97100
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/97832
Tested-by: Szymon Kłos <szymon.klos@collabora.com>

+++ include/LibreOfficeKit/LibreOfficeKit.hxx
@@ -308,7 +308,7 @@ public:
      * @param nWindowId id of the window to notify
      * @param pArguments arguments of the event.
      */
-    void sendDialogEvent(unsigned nWindowId, const char* pArguments = NULL)
+    void sendDialogEvent(unsigned long long int nWindowId, const char* pArguments = NULL)
     {
         mpDoc->pClass->sendDialogEvent(mpDoc, nWindowId, pArguments);
     }
